nonspec: Create dedicated current activities page (#1268)
Following our ongoing discussions about questions of SLSA's current
status, I quickly put together a new page for the website that outlines
our current activities. This is mostly meant to be a starting point that
we can improve upon as we figure out the best way to communicate our
ongoing activities.

---------

Signed-off-by: Marcela Melara <marcela.melara@intel.com>
Co-authored-by: Tom Hennen <TomHennen@users.noreply.github.com>
Co-authored-by: Arnaud J Le Hors <lehors@us.ibm.com>