nonspec: New navigation (#1275)
Here is my first take on the new navigation approach.

This PR doesn't make any changes to how the specification is organized
with regard to the tracks, etc. It is meant to be merely a UI
modification, essentially changing the navigation bar so that the
various versions of the specification are directly visible and
accessible rather than dependent on using the version selector which
only appears once you go into the specification.

I didn't include the older versions: 0.1 and 1.0-rc1 and rc2. All of
these are however still there and can be accessed directly via their
respective URLs or from any page linking to them such as past blog
posts. Let me know if you think we should add 0.1 to the navigation bar.
I'm also not sure there is value in having the 1.1 RC given that I think
it's a dead-end (if anything I think this should be 1.0.1).

In the process of making this change I found a few bugs that I was able
to fix.

Any feedback or suggestions welcome but please keep in mind that Jekyll
is a **static** page generator so anything that requires dynamic
processing (either server-side or client-side) is out of scope. These
would require PHP and/or Javascript. I've more or less managed to get my
head around Jekyll and its template programming language Liquid but I'm
not up for the added complexity any of this would imply.

PR #1268 will have a small impact that I'll handle once it is merged.

---------

Signed-off-by: Arnaud J Le Hors <lehors@us.ibm.com>