Tweak "Compromise project owner" example
The old example was using the term "administrator" when
the threat and mitigation were talking about project
ownership.  This was confusing and could be conflated
with Build _Platfrom_ administration which is very
different.

In addition instead of talking about provenance
forgability (the title of the overall requriement)
it now talks specifically about the need for provenance
to be complete.

Signed-off-by: Tom Hennen <tomhennen@google.com>