Just "Platform Operations Track"
Future directions talks about a 'Platform Operations Track' that
would provide guidance on how to run a secure platform (e.g.
ACLing, logging, secret management).  It _had_ been labeled the
"_Build_ Platform Operations Track" but was referenced from
areas that weren't _build_ platforms.  In fact this guidance
would apply broadly for all platforms.  So, this change removes
the 'build' qualifier.

It remains to be seen when/if this track will be developed.

Signed-off-by: Tom Hennen <tomhennen@google.com>