content: build L3 requires complete external parameters, not provenance
It was raised in the review of 1.1-RC2 that build L3 does not require
provenance to be complete. Instead, it only requires the external
parameters to be complete. With this requirement, it is still sufficient
to mitigate an attack by SSH on the artifact as the verifier would be
able to ensure that the external parameters meet expectations.

ref: https://github.com/slsa-framework/slsa/pull/1298#discussion_r2018863889

Signed-off-by: arewm <arewm@users.noreply.github.com>