Introduce Source Level 4 - Two Party Review
After a great deal of discussion, looking at slsa.dev/threats,
etc, it appears that two-party review is the strongest control
we have against many of the threats listed for 'B'.

This PR introduced a 4th source level which requires two-party
review of changes to protected branches.

As noted by some this is one of the first controls enterprises
enable, while being a control that can be very difficult for
small projects to enable.  To account for this we are making
this the highest source level as 1-3 are much more easily
attained by single-maintainer projects. Those projects
can advance as far as possible without adopting two-party
review.

To further reduce the burden of this requirement we suggest
that reviews cover 'security relevant properties' to allow
reviewers to focus on the most pressing aspects of code-review
and avoid the perception that these reviews require discussion
of 'trivial' issues such as variable names. Of course,
organizations may still set a higher bar for review if they
wish.

Signed-off-by: Tom Hennen <tomhennen@google.com>