content: draft: Add onboarding blurb, clarify 'branch' definition. (#1351)
The content of a revision is built over time. The controls in place
today may be more strict than the controls in place yesterday.

Users of existing repositories will need to be able to onboard those
repositories with some understanding of
what claims are being made about prior revisions.

This is a first attempt at handling that issue.

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>