content: draft: Define the process for verifying source.
This borrows _heavily_ from https://slsa.dev/spec/draft/verifying-artifacts
attempting make changes where needed.

While we might think there is enough overlap that the two could share
the same content, I think that's unlikely. There are enough cases
where details differ that it would likely be more confusing.

This change DOES NOT include "Resulting threat mitigation" which
is included in 'Step 1' of 'verifying-artifacts'.  We may want
to include that in the future or we might prefer to leave it
to the threats.md page.

refs #1356

Signed-off-by: Tom Hennen <tomhennen@google.com>