nonspec: Add some documentation about the website internals (#1360)
I had a first brain dump about some of the internals of slsa.dev that
ought to be better documented because anyone trying to publish a new
version of the spec will need to know about these.

I still plan to do some additional investigation and update as necessary
but I thought it'd be better to share it earlier rather than later. I
invite anyone who's interested to have a look and ask any questions you
may have. This will only help me add whatever else is missing and might
be helpful in the future.

---------

Signed-off-by: Arnaud Le Hors <lehors@us.ibm.com>