editorial: Remove mention of conformance program (#1371)
The 'conformance program' is no longer 'coming soon' so we shouldn't say
that it is.

Updating both 1.1 (current) and draft.

As this doesn't change the meaning of the spec I've listed this as
'editorial' and I don't think we need (or should) increment the 1.1
version number. If anyone thinks otherwise (or that we should just not
update 1.1) LMK.

fixes #1364

Signed-off-by: Tom Hennen <tomhennen@google.com>