content: draft: Merge change mgmt tool requirements with SCS (#1370)
1. Removes the 'Verified timestamp' requirement as the SLSA Source track
doesn't seem to be the right place to address this.
2. Moves the 'Context' requirement up to the "Source Control System"
section.
3. Removes the "Branches may have differing security postures, and a
change can be approved for one context while being unapproved for
another." statement as it's a) explanation rather than requirement, b)
covered by other requirements that talk about how some branches need to
be protected and others don't.
4. Updated the 'Context' requirement to apply at Level 2 (instead of
starting at Level 3) as I believe this is a fairly basic feature of
VCSes. AIUI this requirement isn't about storing this data in
provenance, just making it available. If it _is_ about what gets encoded
in provenance, then we can move it there. :)

It's possible the language can use some additional tweaks for clarity,
but I think it's better to address those separately.

fixes #1352

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>
Signed-off-by: Tom Hennen <TomHennen@users.noreply.github.com>
Co-authored-by: Aditya Sirish <8928778+adityasaky@users.noreply.github.com>
Co-authored-by: Zachariah Cox <zachariahcox@github.com>