fix: remove reference to previous version being level 0.1 (#1359)
This pull request removes outdated explanatory notes from the
`levels.md` documentation file to streamline and clarify the content.

I think we should remove this reference to 0.1 being the "previous
version."
From context it may be that "previous version" is currently encoded
somewhere to always be 0.1.
Edit: found it and deleted it.

For reference, I was poking around here:
https://slsa.dev/spec/v1.1/levels

It has been a full release, so I think it is also fine to remove the
entire note.

---------

Signed-off-by: Zachariah Cox <zachariahcox@gmail.com>