content: draft: Remove MFA requirement
This requirement isn't really source specific. It would apply
to any sort of service offering.  The concepts also don't
map cleanly to potential distributed implementions like
gittuf (where users don't login).

If we want to address this sort of thing in the future we
can include it in a Platform Operations Track
https://slsa.dev/spec/draft/future-directions#platform-operations-track.

fixes #1266

Signed-off-by: Tom Hennen <tomhennen@google.com>