content: draft: MUST NOT be possible -> SCS MUST prevent (#1384)
We have a couple of requirements in the source track that state "It MUST
not be possible...". A strict reader might complain that even if there
are protections in place that prevent this, it might still "be
possible".

To avoid such complaints, and to make it somewhat more clear that these
things could still happen, let's try rephrasing these to something like
"The SCS MUST prevent...".

That should offer a bit more leeway.

fixes #1383

Signed-off-by: Tom Hennen <tomhennen@google.com>