content: draft: further refinement for revision continuity (#1381)
Continuity is relevant for any requirements imposed on successive
revisions within a protected branch. This PR is clarifying that
continuity is established from a given revision and what properties also
adhere to the continuity requirements.

This PR superseded the work in #1349 due to the amount of changes to the
spec that have happened since it was opened.

Signed-off-by: arewm <arewm@users.noreply.github.com>