content: draft: Move "Enforced change management process" to L2
fixes #1391

L2 already covers the requirement that controls need to be in place.
That we need to have continuity for those controls.
That organizations need to specify what things are protected.

Moving the requirement that the SCS enforces those controls to L2
makes the story cleaner (L3 == provenance, L2 == controls) and
should not be a burden.

Updated the threat that referred to this control to be addressed
at L2.

Signed-off-by: Tom Hennen <tomhennen@google.com>