content: draft: Add 'SLSA Verified Properties'
While SLSA is typically focused on expressing the security state of an artifact
with [levels](principles#simple-levels-with-clear-outcomes), levels may not be
appropriate in all cases. Some software supply chain controls don't fit neatly
within existing SLSA levels or that do may exist with others that users are not
yet able to meet.  SLSA Verified Properties allows a common way to express those
properties (and their requirements) without needing to fit them into existing
levels or introduce new tracks.

These properties MAY be included in the `verifiedLevels` field of
[verification_summaries (VSAs)](verification_summary) when the VSA issuer
determines the requirements have been met.

Note that this also introduces a new `SLSA_BUILD_REPRODUCED` property that
could be useful in resolving #382 but _could also be controversial_. We do not
need to keep it if it's too difficult to get agreement (e.g. we can defer that
discussion for later if we want).

Also renames provenance.md/requirements.md to build-provenance.md/build-requirements.md
to distinguish them from the source track requirements.  Maybe this also is too
big of a change.  I can certainly revert it if it's easier.

fixes #1355

Signed-off-by: Tom Hennen <tomhennen@google.com>