content: draft: Consolidate control requirements (#1395)
fixes #1391
fixes #1390 

## Move "Enforced change management process" to L2

L2 already covers
* That controls need to be in place.
* That we need to have continuity for those controls.
* That organizations need to specify what things are protected.

Moving the requirement that the SCS enforces those controls to L2 makes
the story cleaner (L3 == provenance, L2 == controls) and should not be a
burden.

Updated the threat that referred to this control to be addressed at L2.

## Combine choose-process, specify-control-expectations, and
specify-protection

Now that all of these things occur at L2 it makes it possible to combine
them into one requirement. This mirrors the Build Track which has
multiple requirements in each row. I think it also winds up making it
more clear
how the requirements relate to one another.

Also made all the example text 'quotes' to better separate it from the
requirements themselves.

## Require ACLs and orgs to use them.

Added a requiremen to 'Identity Management' that the IMS provide the
ability to specify which actors/roles can perform protected actions.

Also requiring the organization to use the IMS to configure those
actors/roles.

## Update L2 summary

Level 2 is now about a lot more than just tracking changes and ensuring
we can attribute them.
It is much more 'control' focused. Updating the summary to capture that.

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>
Signed-off-by: Tom Hennen <TomHennen@users.noreply.github.com>
Co-authored-by: Zachariah Cox <zachariahcox@github.com>