editorial: Update principles.md to align verbiage (#1404)
Aligning the verbiage, so that we consistently refer to users as
'individuals' rather than 'people'.

Per https://github.com/slsa-framework/slsa/issues/1376, reviewing uses
of the term's 'people' and 'person'
In the principles section, we headline with the term 'individuals', but
later refer to them as 'people'. This change replaces 'people' with
'individuals.

Signed-off-by: Tom Bedford <56314245+tombedfordgit@users.noreply.github.com>