content: draft: Move Common Source Controls to Requirements (#1398)
Moving "Common Source Controls and their Applications" from
verifying-source.md to source-requirements.md.

In the process I rejiggered the introduction to focus on the control
aspect and how they can be validated with some of the information in the
VSA or provenance (based on the enforced change management process
stuff).

I also cut the two party review control because that's now a part of the
levels themselves so it shouldn't need mentioning.

It might be worth linking to the verifying-source page at the start??

fixes #1396

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>