content: draft: Update VSA spec to handle SLSA levels beyond the build track (#1417)
This basically follows the same strategy as before but provides a
formula
instead of a full enumeration and links to verified-properties.md for
additional SLSA_ values.

I've decided _not_ to include a prohibition on `ORG_` prefixed policies
because
the SlsaResult documentation already tells users (i.e. VSA issuers) that
they
MAY use custom values as long as they don't start with `SLSA_`. `ORG_`
prefixed values, as they're defined by the organization, _are_ custom
values
so it seems we're already aligned with the spec. We're also not trying
to
establish (within SLSA) any specific meaning for these ORG properties.
That's left to the org itself to document.

fixes https://github.com/slsa-framework/slsa/issues/1369

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>