content: draft: 'Context' -> 'History' (#1416)
Reworded the 'Context' requirement to a 'History' requirement
taking some content from the 'Continuity' requirement to do so.

The continuity requirement now focuses on the need to maintain
continuity of controls (both history and enforced change
management process) and leaves much of the rest of the text
to 'history'.

Moved 'protected branches' ahead of  'history' since history
only applies to protected branches (so seems like a better
ordering).

fixes #1407

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>
Signed-off-by: Tom Hennen <TomHennen@users.noreply.github.com>
Co-authored-by: Marcela Melara <marcela.melara@intel.com>
Co-authored-by: Aditya Sirish <8928778+adityasaky@users.noreply.github.com>