content: Create SLSA 1.2-rc1 (#1428)
Creates SLSA 1.2-rc1

* I copied docs/spec/draft to docs/spec/v1.2-rc1
* I removed references to the build environment track from within
v1.2-rc1
* I removed references to verified-properties from within v1.2-rc1
* I then followed [these
instructions](https://github.com/slsa-framework/slsa/blob/main/docs/README.md#publishing-a-new-version-of-the-slsa-specification)
to update nav, etc.
* I did _not_ create a layout/v1.2-rc1.yml because according to [these
instructions](https://github.com/slsa-framework/slsa/blob/main/docs/README.md#publishing-a-new-version-of-the-slsa-specification)
that should only be needed if the version is hidden. Since this isn't
it's redundant to have it listed in main and in it's own layout (and
likely error prone?).

fixes #1419

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>
Signed-off-by: Tom Hennen <TomHennen@users.noreply.github.com>
Co-authored-by: Zachariah Cox <zachariahcox@github.com>