content: draft: Human Readable Changes (#1472)
fixes #1460

Add a requirement at L1 that SCS's provide tooling to display changes in
human readable form. Used most of the langage from 'informed review'
(L4) and then having 'informed review' reference this requirement.

Note that this is a MUST for plain text files but a SHOULD for non-plain
text.

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>
Signed-off-by: Tom Hennen <TomHennen@users.noreply.github.com>
Co-authored-by: Zachariah Cox <zachariahcox@github.com>