content: draft: Update Source Track terms (#1471)
fixes #1461, which has a long discussion on the topic.

* Removed 'Source' using 'Source Repository' instead.
* Renamed 'Version Control System' to 'Source Management Tool'. It's
helpful to have a term to refer to the tools since they're so common
place and folks will naturally wonder where specific tools (e.g. git)
fit into this framework.
* Updated 'Source Control System' definition to largely match what was
discussed in the issue.
* Updated 'Source Repository' definition, based on what was discussed in
the issue but with more tweaks.
* Updated 'Source Revision' definition to largely match what was
discussed in the issue.
* Updated 'Branch' definition based on the issue.
* Updated 'Tag' defintiion based on text in the issue but tweaked to
allow for mutable tags (e.g. 'latest', 'nightly').
* Updated 'Change' based on what was discussed in the issue.

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>
Signed-off-by: Tom Hennen <TomHennen@users.noreply.github.com>
Signed-off-by: Zachariah Cox <zachariahcox@github.com>
Co-authored-by: Zachariah Cox <zachariahcox@github.com>
Co-authored-by: Mark Lodato <lodatom@gmail.com>