editorial: draft: Move example controls to its own page. (#1503)
Move the example controls from source-requirements.md to their own page
and improve their formatting.

These examples were never meant to be a part of the spec itself but
rather as example controls that others may want to implement or that
would serve as inspiration for developing their own controls. They go
above and beyond the specific requirements of the source track itself.

This change:

1. Moves the content to its own page, adding links to the nav bar and
tracks.md page.
2. Improves the formatting of the content, making summary, intended for,
and benefits a table to make it easier to parse.
3. Fills in some gaps that existed: one section didn't have a summary,
etc, another was missing the 'intended for' information.
4. Updated the introduction so that it doesn't reference SLSA L4. These
examples can be used as a part of L4 but they don't need to be (e.g. the
testing or merge train examples).

Points 3 and 4 required changing this to an 'editorial' fix instead of a
'style' fix.

fixes #1465

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>