content: draft: Remove provenance requirement
Creating this PR to get an idea of what it would mean to remove
the provenance (and VSA) requirement from the spec.

It seems like the majority of what we're looking for in provenance
can be subsumed by other existing requriements we already track
with provenance becoming a (convenient) implementation detail.

If we do this we might also want to move the provenance and VSA
parts of this page to their own page to keep the level requirements
clear.

If we accept this change it would mean that 'provenance' becomes
a nice feature of any particular SLSA Source Track implementation
but won't be required.

Signed-off-by: Tom Hennen <tomhennen@google.com>