editorial: draft: Note the purpose of the provenance build model diagram
We have two diagrams that are very similar and both depict the build model.

This one, which is specific to the provenance format, and the one on the
https://slsa.dev/spec/draft/terminology#build-model page which is less
specific.  We probably want both versions.  So, instead of updating the
diagrams I've added an explanatory note with a link to prevent confusion.

fixes #1290

Signed-off-by: Tom Hennen <tomhennen@google.com>